% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idftags.R
\name{build_idf_tag}
\alias{build_idf_tag}
\title{Build EnergyPlus IDF Tag file}
\usage{
build_idf_tag(path, out = "|")
}
\arguments{
\item{path}{\code{\link[=character]{character()}} A string specifying the path of an EnergyPlus IDF
file.}

\item{out}{\code{\link[=character]{character()}} The output tag file path. If \code{"|"}, the tag file
will be directly printed into stdout. Default: \code{"|"}}
}
\description{
\code{build_idf_tag()} extracts the locations of classes and objects in the
input EnergyPlus IDF and creates a
\href{https://github.com/universal-ctags/ctags}{ctags}-compatible tag file.
}
\examples{
\dontrun{
if (requireNamespace("eplusr", quietly = TRUE)) {
build_idf_tag(system.file("extdata/1ZoneUncontrolled.idf", package = "eplusr"))
}
}
}
