% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idftags.R
\name{build_idf_tags}
\alias{build_idf_tags}
\title{Build EnergyPlus IDF Tag file}
\usage{
build_idf_tags(path, out = "|", cmd = FALSE)
}
\arguments{
\item{path}{\code{\link[=character]{character()}} A string specifying the path of an EnergyPlus IDF
file.}

\item{out}{\code{\link[=character]{character()}} The output tag file path. If \code{"|"}, the tag file
will be directly printed into stdout. Default: \code{"|"}.}

\item{cmd}{\code{\link[=logical]{logical()}} If \code{TRUE} and both \code{path} and \code{out} are missing, try
to read arguments using \code{\link[=commandArgs]{commandArgs()}} for \code{path} and \code{out}. This enables
you to use \code{build_idf_tags()} in the way like
\verb{Rscript -e "idftags:: build_idf_tags(cmd = TRUE)" path}. Default: \code{FALSE}.}
}
\description{
\code{build_idf_tags()} extracts the locations of classes and objects in the
input EnergyPlus IDF and creates a
\href{https://github.com/universal-ctags/ctags}{ctags}-compatible tag file.
}
\examples{
\dontrun{
if (requireNamespace("eplusr", quietly = TRUE)) {
build_idf_tags(system.file("extdata/1ZoneUncontrolled.idf", package = "eplusr"))
}
}
}
