% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idftags.R
\name{read_object_location}
\alias{read_object_location}
\title{Read object locations}
\usage{
read_object_location(path)
}
\arguments{
\item{path}{\code{\link[=character]{character()}} A string specifying the path of an EnergyPlus IDF
file.}
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}} with 7 columns:
\itemize{
\item \code{class_id}: \code{\link[=integer]{integer()}} Integers as the identifiers for each class detected.
\item \code{class_line}: \code{\link[=integer]{integer()}} Line numbers indicating where the classes are found
\item \code{class_name}: \code{\link[=character]{character()}} Standard EnergyPlus class names
\item \code{class_string}: \code{\link[=character]{character()}} The raw input string for classes
\item \code{object_id}: \code{\link[=integer]{integer()}} Integers as the identifiers for each object detected.
\item \code{object_line}: \code{\link[=integer]{integer()}} Line numbers indicating where the objects are found
\item \code{object_name}: \code{\link[=character]{character()}} Object names found
\item \code{object_string}: \code{\link[=character]{character()}} The raw input string for objects
}
}
\description{
\code{read_object_location()} extracts the locations of classes and objects in the
input EnergyPlus IDF.
}
\examples{
\dontrun{
if (requireNamespace("eplusr", quietly = TRUE)) {
read_object_location(system.file("extdata/1ZoneUncontrolled.idf", package = "eplusr"))
}
}
}
